A: Device. #_start_app. We need to port forward.
Imagine this scenario:
Port H is forwarded to port A.
A test-server is running on port A.
The port changes to B.
\#start_app is called.
\#start_app will try to close down any old test-server.
As H is currently forwarded to A, we can see a test-server running.
But we should not close down A. Therefore we forward port H to B,
before doing the check.