Defaults: Add default methods should be static methods
One of the reasons is the following:

If the instance method m and m= are defined in the scope,
and m returns an object that has the method foo, we would
expect
`m = m.foo`
to be equivalent to
`m=(m().foo)`
but it is not. Instead it would raise an error saying that
m does not have the method foo, as m is instead referencing a local
variable, as m = is evaluated differently.