iOS: Don't always restart sims in start_app
This change is a bit controversial. We do not ask RunLoop to quit the
simulator when we want to start an application on it.

Note not RunLoop will *still* restart the sim if it determines that it
is not stable. I tried launching the sim from XCode, and RunLoop did
restart it.